+++ example/lib/pages/input_form_field.dart
@@ -68,6 +68,10 @@ class _InputFormFieldPageState extends State<InputFormFieldPage> {
               children: [
                 ShadInputFormField(
                   id: 'username',
+                  prefix: const ShadImage.square(
+                    size: 16,
+                    LucideIcons.user,
+                  ),
                   label: const Text('Username'),
                   placeholder: const Text('Enter your username'),
                   description: const Text('This is your public display name.'),