+++ lib/src/components/input.dart
@@ -7,6 +7,7 @@ import 'package:flutter/services.dart';
 import 'package:shadcn_ui/src/components/disabled.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
+import 'package:shadcn_ui/src/utils/separated_iterable.dart';
 
 class ShadInput extends StatefulWidget {
   const ShadInput({
@@ -80,6 +81,7 @@ class ShadInput extends StatefulWidget {
     this.placeholderStyle,
     this.placeholderAlignment,
     this.inputPadding,
+    this.gap,
   })  : smartDashesType = smartDashesType ??
             (obscureText ? SmartDashesType.disabled : SmartDashesType.enabled),
         smartQuotesType = smartQuotesType ??
@@ -162,6 +164,7 @@ class ShadInput extends StatefulWidget {
   final Alignment? placeholderAlignment;
   final EdgeInsets? inputPadding;
   final bool onPressedAlwaysCalled;
+  final double? gap;
 
   static const int noMaxLength = -1;
 
@@ -185,7 +188,9 @@ class ShadInputState extends State<ShadInput>
   @override
   void initState() {
     super.initState();
-    if (widget.focusNode == null) _focusNode = FocusNode();
+    if (widget.focusNode == null) {
+      _focusNode = FocusNode(canRequestFocus: !widget.readOnly);
+    }
     focusNode.addListener(onFocusChange);
 
     if (widget.controller == null) {
@@ -206,6 +211,10 @@ class ShadInputState extends State<ShadInput>
       unregisterFromRestoration(_controller!);
       _controller!.dispose();
       _controller = null;
+    }
+
+    if (widget.readOnly != oldWidget.readOnly) {
+      _focusNode?.canRequestFocus = widget.readOnly;
     }
   }
 
@@ -299,6 +308,8 @@ class ShadInputState extends State<ShadInput>
         CrossAxisAlignment.center;
     final effectiveMouseCursor =
         widget.mouseCursor ?? WidgetStateMouseCursor.textable;
+
+    final effectiveGap = widget.gap ?? theme.inputTheme.gap ?? 8.0;
 
     final defaultSelectionControls = switch (Theme.of(context).platform) {
       TargetPlatform.iOS => cupertinoTextSelectionHandleControls,
@@ -444,7 +455,7 @@ class ShadInputState extends State<ShadInput>
                           ),
                         ),
                         if (widget.suffix != null) widget.suffix!,
-                      ],
+                      ].separatedBy(SizedBox(width: effectiveGap)),
                     ),
                   ),
                 );